loader is now a wrapper component
