+++ src/App.css
@@ -74,7 +74,7 @@ body {
 	align-items: center;
 	justify-content: space-between;
 	margin: 0;
-	padding: 0 10px 10px 25px;
+	padding: 5px 10px 10px 35px;
 	width: 100%;
 	grid-column: 2 / -1;
 	grid-row: 1 / 2;
@@ -126,11 +126,18 @@ body {
 	padding: 15px;
 }
 
-.loadingRef {
+/* .loadingRef {
 	position: relative;
 	width: 100%;
 	margin: 0;
 	padding: 0;
+	grid-column: 2 / 4;
+	grid-row: 1 / 2;
+} */
+
+.LineGraph {
+	width: 100%;
+	margin: 0;
 	grid-column: 2 / 4;
 	grid-row: 1 / 2;
 }
@@ -140,11 +147,6 @@ body {
 	grid-row: 3 / 4;
 	width: 100%;
 	margin: 0;
-}
-
-.LineGraph {
-	/* max-width: 100%; */
-	/* max-height: 100%; */
 }
 
 .DoughnutGraph {
@@ -185,7 +187,7 @@ body {
 	.totalLineGraph {
 		grid-column: 1 / 4;
 	}
-	.loadingRef {
+	.LineGraph {
 		grid-column: 1 / 4;
 	}
 	.leftSubGrid {
@@ -520,7 +522,7 @@ label[for="companySelect"] {
 	margin: 2px;
 }
 
-#companySelect {
+.CompanySelect select {
 	margin-top: 0;
 	font-size: 14px;
 	padding: 0.1rem;
@@ -529,24 +531,25 @@ label[for="companySelect"] {
 	width: 100%;
 	border: none;
 }
-#companySelect:hover {
+.CcompanySelect select:hover {
 	background-color: rgb(0,255,255);
 	transition-duration: 0.2s;
 }
-#companySelect:disabled {
-	margin-top: 0;
-	font-size: 14px;
-	padding: 0.1rem;
+.CompanySelect select:disabled {
 	background-color: rgb(113, 57, 14);
-	text-align: center;
-	width: 100%;
-	border: none;
+	max-width: 100%;
+	min-width: 266px;
 }
 ::placeholder {
 	color: black;
 }
 /* =========================waiting for database indicator=========================== */
-
+.loadingRef {
+	position: relative;
+	width: 100%;
+	margin: 0;
+	padding: 0;
+}
 .loadingFlex {
 	position: absolute;
 	top:0;
@@ -569,5 +572,6 @@ label[for="companySelect"] {
 	border: 2px solid rgb(240,240,0);
 	border-radius: 5px;
 	background: black;
+	z-index: 3;
 }
